--- backend_theme_v12/models/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import res_company
-from . import res_users
