--- backend_theme_v12/views/web.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-    <template id="menu" inherit_id="web_responsive.apps">
-    <!--    <xpath expr="//div[hasclass('panel-heading')]"
-            position="replace" /> -->
-        
-        <xpath expr="//div[hasclass('dropdown-menu')]" position="after">
-                <div class="panel-footer" id="appDrawerAppPanelFooter">
-                    <div class="oe_logo">
-                        <img src="/web/binary/company_logo" />
-                    </div>
-                </div>
-        </xpath>
-    </template>
-
-    <!-- <template id="webclient_bootstrap"
-        inherit_id="web_responsive.webclient_bootstrap">
-        <xpath expr="//i[hasclass('fa-th')]" position="after">
-            <i class="fa fa-lg fa-chevron-left app-drawer-icon-close"
-                t-translation="off" aria-hidden="true" />
-        </xpath>
-    </template> -->
-
-</odoo>
