+++ backend_theme_v13/models/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+from . import res_company
+#from . import res_users
