+++ backend_theme_v13/static/description/index.html
@@ -1,0 +1,45 @@
+<section class="oe_container">
+    <div class="oe_row oe_spaced">
+		<div class="oe_span12">
+			<h1 class="oe_slogan" style="color:#337ab7;">Openworx Material Backend Theme</h1>
+			<h2 class="oe_slogan">The most popular backend theme for Odoo 13.0 Community</h2>
+			<div class="oe_screenshot" style="text-align: center;">
+				<h3 class="oe_slogan" style="color:#337ab7;">Enterprise look'n'feel App dashboard</h3>
+				<img style="max-width: 100%;" src="screenshot1.png"/><br><br>
+				<h3 class="oe_slogan" style="color:#337ab7;">Formview</h3>
+				<img style="max-width: 100%;" src="screenshot2.png"/><br><br>
+			</div>
+		</div>
+	</div>
+</section>
+
+<section class="oe_container oe_separator">
+</section>
+
+<section class="oe_container" style="background-color:#E8E8E8; padding:20px 0px 90px 0px">
+    <div class="oe_row">
+        <div class="oe_span12">
+            <div class="oe_centeralign">
+                <p>You can change the App dashboard background under Company Settings.</p>
+                <p>Remark: It is possible that the Odoo service needs to be restarted after install.</p>
+            </div>
+        </div>	
+    </div>
+</section>
+
+<section class="oe_container oe_separator">
+</section>
+
+<section class="oe_container oe_dark">	
+    <div class="oe_row">	
+        <div class="oe_span12 oe_centeralign">	
+            <br/>	
+            <h3>Help & Support</h3>	
+            <br/>	
+                Website: <a href="https://www.openworx.nl" target="_blank">Openworx</a><br/>	
+                Contact: <a href="mailto:info@openworx.nl" subject="Material Backend Theme v12">Email Support</a><br/>	
+            <br/>	
+        </div>	
+    </div>	
+</section>
+
