+++ backend_theme_v13/static/src/scss/sidebar.scss
