--- openworx_backend_theme/README.rst
@@ -1,1 +1,0 @@
-This module has been replaced by https://github.com/Openworx/backend_theme/tree/12.0/backend_theme_v12
